Move product data source into products package
